requirements.txt
Flask==3.0.0
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
Flask-Session==0.5.0
gunicorn==21.2.0
